requirements.txt
google-analytics-admin>=0.28.0,<1.0.0
google-auth>=2.37.0,<3.0.0
google-auth-oauthlib>=1.3.0,<2.0.0
google-api-python-client>=2.155.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
